requirements.txt

Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-yasg==1.20.0

gunicorn==20.1.0

Jinja2==3.1.2
keyring==18.0.1

requests==2.22.0

sqlparse==0.4.2

whitenoise==6.0.0

